Revert "Added a new setting which disable docker version validation."
This reverts commit 4d612c67
